agents/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spaceapss-agents"
version = "0.1.0"
description = "Agente de busca e resposta sobre artigos científicos"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SpaceAPSS Team"}
]
dependencies = [
    # API e Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # HTTP e Scraping
    "aiohttp>=3.9.5",
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.3",
    "lxml[html_clean]>=5.3.0",
    "readability-lxml>=0.8.1",
    
    # PDF Processing
    "pdfminer.six>=20240706",
    
    # Azure
    "azure-functions>=1.21.3",
    "azure-cosmos>=4.5.0",
    
    # Database
    "redis>=5.0.0",
    "pymongo>=4.6.0",
    "motor>=3.3.0",
    "chromadb>=0.4.0",
    
    # AI/ML
    "google-generativeai>=0.3.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pandas>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "ipython>=8.20.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["packages*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "packages/api/tests",
]
pythonpath = ["."]
asyncio_mode = "auto"